restructured pkg/progutils
