+++ pkg/progutils/utils.go
@@ -19,30 +19,14 @@ package progutils
 */
 
 import (
-	"fmt"
 	"log/slog"
-	"math"
-	"sort"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/charmbracelet/lipgloss"
-	"github.com/charmbracelet/lipgloss/table"
 	"github.com/charmbracelet/log"
 	"github.com/sethll/goCBC/pkg/hlcalc"
-	"github.com/sethll/goCBC/pkg/progmeta"
 )
-
-// StylesType defines the styling configuration for different UI elements.
-type StylesType struct {
-	Bedtime      lipgloss.Style
-	Caffeine     lipgloss.Style
-	Header       lipgloss.Style
-	TableHeader  lipgloss.Style
-	TableEvenRow lipgloss.Style
-	TableOddRow  lipgloss.Style
-}
 
 // TimeAndAmount represents a time entry with an associated substance amount.
 type TimeAndAmount struct {
@@ -51,37 +35,7 @@ type TimeAndAmount struct {
 	TimeObject time.Time
 }
 
-type Results struct {
-	ChemIngestedTotal            float64
-	BodyChemContent              float64
-	TheoreticalChemIngestedTotal float64
-	TheoreticalBodyChemContent   float64
-	LastRealTime                 time.Time
-	WearoffTime                  time.Time
-	TheoreticalWearoffTime       time.Time
-}
-
 var (
-	// Styles contains the styling configuration for different UI elements.
-	Styles = StylesType{
-		Bedtime: lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#87CEEB")),
-		Caffeine: lipgloss.NewStyle().
-			Foreground(lipgloss.Color("#FFA500")),
-		Header: lipgloss.NewStyle().
-			Bold(true).
-			Padding(1, 2).
-			Border(lipgloss.RoundedBorder()),
-		TableHeader: lipgloss.NewStyle().
-			Bold(true).
-			Padding(0, 2),
-		TableEvenRow: lipgloss.NewStyle().
-			Padding(0, 2),
-		TableOddRow: lipgloss.NewStyle().
-			Padding(0, 2).
-			Background(lipgloss.Color("#888888")),
-	}
-
 	// LogLevelSelector maps verbosity levels to log levels.
 	LogLevelSelector = map[int]log.Level{
 		0: log.ErrorLevel,
@@ -90,41 +44,6 @@ var (
 		3: log.DebugLevel,
 	}
 )
-
-// GenerateOutputTable creates a formatted table displaying current substance levels
-// and anticipated bedtime based on the target amount.
-func GenerateOutputTable(chemInBody *float64, bedTime *time.Time, sleepTarget *string, chem *string) *table.Table {
-	slog.Debug("Generating output table", "chemInBody", *chemInBody, "bedTime", (*bedTime).Format("2006-01-02 15:04"), "sleepTarget", *sleepTarget)
-	rows := [][]string{
-		{
-			fmt.Sprintf(
-				"%s remaining in system:",
-				Styles.Caffeine.Render(StringToTitleCase(*chem)),
-			),
-			Styles.Caffeine.Render(
-				fmt.Sprintf(
-					"~%.0fmg",
-					math.Round(*chemInBody),
-				),
-			),
-		},
-		{
-			fmt.Sprintf(
-				"Reach target (%s) for %s at:",
-				Styles.Caffeine.Render(
-					fmt.Sprintf("%smg", *sleepTarget),
-				),
-				Styles.Bedtime.Render("sleep"),
-			),
-			Styles.Bedtime.Render(
-				(*bedTime).Format("2006-01-02 15:04"),
-			),
-		},
-	}
-	generatedTable := table.New().Border(lipgloss.HiddenBorder()).Rows(rows...)
-	slog.Debug("Output table generated successfully", "rowCount", len(rows))
-	return generatedTable
-}
 
 // GetTimesAndAmounts parses command-line time:amount strings into TimeAndAmount structs.
 // Input format should be "HHMM:amount" (e.g., "1100:300").
@@ -153,86 +72,6 @@ func GetTimesAndAmounts(inputs *[]string) (returnList []TimeAndAmount) {
 	returnList = sortTimeEntries(returnList)
 	slog.Debug("Completed processing inputs", "validEntries", len(returnList), "totalInputs", len(*inputs))
 	return
-}
-
-// NewResults initializes a new Results
-func NewResults() Results {
-	return Results{
-		ChemIngestedTotal:            0,
-		BodyChemContent:              0,
-		TheoreticalChemIngestedTotal: 0,
-		TheoreticalBodyChemContent:   0,
-		LastRealTime:                 time.Time{},
-		WearoffTime:                  time.Time{},
-		TheoreticalWearoffTime:       time.Time{},
-	}
-}
-
-// PrintProgHeader displays the program header with name, version, and copyright information.
-func PrintProgHeader() {
-	slog.Debug("Printing program header", "progName", progmeta.ProgName, "version", progmeta.AllVersionBuildRuntimeInfo())
-	fmt.Println(
-		Styles.Header.Render(
-			fmt.Sprintf(
-				"%s\t© %s %s\n%s",
-				progmeta.ProgName,
-				progmeta.CopyrightYear,
-				progmeta.Author,
-				progmeta.AllVersionBuildRuntimeInfo(),
-			),
-		),
-	)
-}
-
-func (r *Results) String() string {
-	return fmt.Sprintf("Results{ChemIngestedTotal: %.2f, BodyChemContent: %.2f, TheoreticalChemIngestedTotal: %.2f, TheoreticalBodyChemContent: %.2f, LastRealTime: %s, WearoffTime: %s, TheoreticalWearoffTime: %s}",
-		r.ChemIngestedTotal,
-		r.BodyChemContent,
-		r.TheoreticalChemIngestedTotal,
-		r.TheoreticalBodyChemContent,
-		r.LastRealTime.Format("2006-01-02 15:04:05"),
-		r.WearoffTime.Format("2006-01-02 15:04:05"),
-		r.TheoreticalWearoffTime.Format("2006-01-02 15:04:05"))
-}
-
-func (r *Results) getChemIngestedTotal(theoreticalMode bool) float64 {
-	if theoreticalMode {
-		slog.Debug("Returning theoretical", "TCIT", r.TheoreticalChemIngestedTotal)
-		return r.TheoreticalChemIngestedTotal
-	} else {
-		slog.Debug("Returning real", "CIT", r.ChemIngestedTotal)
-		return r.ChemIngestedTotal
-	}
-}
-
-func (r *Results) setChemIngestedTotal(theoreticalMode bool, input float64) {
-	if theoreticalMode {
-		r.TheoreticalChemIngestedTotal = input
-		slog.Debug("Setting theoretical", "input", input, "TCIT", r.TheoreticalChemIngestedTotal)
-	} else {
-		r.ChemIngestedTotal = input
-		slog.Debug("Setting real", "input", input, "CIT", r.ChemIngestedTotal)
-	}
-}
-
-func (r *Results) getBodyChemContent(theoreticalMode bool) float64 {
-	if theoreticalMode {
-		slog.Debug("Returning theoretical", "TBCC", r.TheoreticalBodyChemContent)
-		return r.TheoreticalBodyChemContent
-	} else {
-		slog.Debug("Returning theoretical", "BCC", r.BodyChemContent)
-		return r.BodyChemContent
-	}
-}
-
-func (r *Results) setBodyChemContent(theoreticalMode bool, input float64) {
-	if theoreticalMode {
-		r.TheoreticalBodyChemContent = input
-		slog.Debug("Setting theoretical", "input", input, "TBCC", r.TheoreticalBodyChemContent)
-	} else {
-		r.BodyChemContent = input
-		slog.Debug("Setting real", "input", input, "BCC", r.BodyChemContent)
-	}
 }
 
 // RunHLCalculations performs half-life calculations for multiple substance intakes
@@ -351,16 +190,6 @@ func RunHLCalculations(results *Results, timesAndAmounts *[]TimeAndAmount, targe
 		"finalTheoreticalBodyChemContent", (*results).TheoreticalBodyChemContent,
 		"wearoffTime", (*results).WearoffTime.Format("2006-01-02 15:04:05"),
 		"theoreticalWearoffTime", (*results).TheoreticalWearoffTime.Format("2006-01-02 15:04:05"))
-}
-
-// sortTimeEntries sorts a slice of TimeAndAmount structs by their TimeString field.
-func sortTimeEntries(timesAndAmounts []TimeAndAmount) []TimeAndAmount {
-	// Sort the slice of TimeAndAmount structs by TimeString
-	sort.Slice(timesAndAmounts, func(i, j int) bool {
-		return timesAndAmounts[i].TimeString < timesAndAmounts[j].TimeString
-	})
-	slog.Info("Sorted timesAndAmounts by TimeAndAmount.TimeString", "timesAndAmounts", timesAndAmounts)
-	return timesAndAmounts
 }
 
 // StringToTitleCase converts a string to title case (first letter uppercase, rest lowercase).
